Finished interface to user and added bubble loss rates.
Now will display all stats and a winning/losing pop-up depending if they won or lost.  Also gives an introduction about the game at the start and from the about menu option (both these pause the game).  The name I came up with is just a placeholder (unless you want to keep it).

As the environment gets dirty, bubble loss rate goes from 0 (clean) to 3 (full dirty) bubbles / second (check a timer >= 1.00 in CAquarium::Update()).

NOTE: Did not test if this is balanced yet since David made the costs for stuff to work so I figured we could balance when we combine the final code.  All values are handled within CAquarium::Update() and displayed in CFrame::Report().